Group 'kopia maintenance' commands in single file
